TWINS-394: adding twin_id to PG* and WT* on DTO level + small fix upon TWINS-395

+++ core/src/main/java/org/twins/face/dao/page/pg002/FacePG002TabEntity.java
@@ -50,7 +50,7 @@ public class FacePG002TabEntity implements EasyLoggable {
     private ResourceEntity iconResource;
 
     @Column(name = "`order`")
-    private Integer order;
+    private short order;
 
     @Transient
     @EqualsAndHashCode.Exclude
